requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ML/AI dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0

# HuggingFace Hub
huggingface_hub>=0.17.0

# Utilities
python-multipart>=0.0.6
requests>=2.28.0
protobuf>=3.20.0
tiktoken>=0.5.0
sentencepiece>=0.1.99
python-dotenv>=1.0.0
psutil>=5.9.0